Pass the render callback code as a pointer-sized integer
GObj_RenderFunc's second argument is not always a pass number: the SIS text renderer is also called directly with an HSD_Text* in it (lb_0192), which truncates on 64-bit hosts. Widen the parameter to intptr_t and give every render callback the matching signature, removing the five function-pointer casts that papered over mismatches (two s32 variants and two one-argument callbacks cast through Event). intptr_t is int under MWCC, so the retail build is unchanged.

+++ src/melee/cm/cmsnap.h
@@ -9,6 +9,6 @@ void cmSnap_800315C8(void);
 void cmSnap_800316B4(void);
 
 void* cmSnap_80031618(void);
-void cmSnap_80031640(HSD_GObj* gobj, int code);
+void cmSnap_80031640(HSD_GObj* gobj, intptr_t code);
 
 #endif
